Separated interfaces.
