+++ examples/basics/src/actors/actor.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 import { CounterActor } from './counter.actor';
-import { GlobalCounterActor } from './global.counter.actor';
+import { GlobalCounterActor } from './global-counter.actor';
 
 export const Actors = [CounterActor, GlobalCounterActor];
