+++ examples/basics/src/actors/counter.actor.interface.ts
@@ -1,0 +1,4 @@
+export abstract class CounterActorInterface {
+  abstract increment(): Promise<number>;
+  abstract getCounter(): Promise<number>;
+}
