+++ examples/basics/src/actors/global-counter.actor.interface.ts
@@ -1,0 +1,4 @@
+export abstract class GlobalCounterActorInterface {
+  abstract increment(): Promise<number>;
+  abstract getCounter(): Promise<number>;
+}
