+++ examples/basics/src/counter/get-counter-value-query.handler.ts
@@ -1,8 +1,8 @@
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { GetCounterValueQuery } from './get-counter-value.query';
 import { DaprActorClient } from '@jeremycarter/nest-dapr';
-import { CounterActorInterface } from '../actors/counter.actor';
-import { GlobalCounterActorInterface } from '../actors/global.counter.actor';
+import { GlobalCounterActorInterface } from '../actors/global-counter.actor.interface';
+import { CounterActorInterface } from '../actors/counter.actor.interface';
 
 @QueryHandler(GetCounterValueQuery)
 export class GetCounterValueQueryHandler
