+++ examples/basics/src/counter/increment-command.handler.ts
@@ -1,9 +1,9 @@
 import { CommandHandler, EventBus, ICommandHandler } from '@nestjs/cqrs';
 import { IncrementCounterCommand } from './increment-counter.command';
 import { DaprActorClient } from '@jeremycarter/nest-dapr';
-import { CounterActorInterface } from '../actors/counter.actor';
 import { CounterIncrementedEvent } from './counter-incremented.event';
 import { Logger } from '@nestjs/common';
+import { CounterActorInterface } from '../actors/counter.actor.interface';
 
 @CommandHandler(IncrementCounterCommand)
 export class IncrementCommandHandler
